backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
django-environ==0.12.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=1.11.0
optimum>=2.0.0
pydub>=0.25.1
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
psutil>=7.1.0
gunicorn>=23.0.0
whitenoise>=6.11.0
dj-database-url>=2.2.0
djangorestframework-simplejwt>=5.3.0
requests>=2.32.0
